Adds the spiderbot fabricator: a machine R&D can build that allows ghosts to respawn as spiderbots (#37375)
* Add the spiderbot fabricator

The spiderbot fabricator is a device built by R&D that can be activated by ghosts to spawn there as a spiderbot.
It is similar to the MoMMI fabricator, except it  must be fed metal sheets to work and can actually be built by players.
It can also be used by mice by clicking on it to become a mouse-spiderbot (which is apparently a thing).

* fix linter error

* Load up with positronic brains instead of metal

* var/i

* shut up this warning too

* Improve this logic

---------

Co-authored-by: west3436 <66280799+west3436@users.noreply.github.com>